Make EconML optional
