+++ .github/workflows/docs.yml
@@ -25,7 +25,7 @@ jobs:
           fetch-depth: 0
 
       - name: Install Python Dependencies
-        run: poetry install -E plotting -E pydot -E pygraphviz --with docs
+        run: poetry install -E plotting -E pydot -E pygraphviz -E econml --with docs
 
       - run: git config --global --add safe.directory /__w/dowhy/dowhy
