+++ .github/workflows/nightly-tests.yml
@@ -35,7 +35,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           echo "installing poetry dependencies"
-          poetry install -E plotting -E pydot -E pygraphviz --with docs
+          poetry install -E plotting -E pydot -E pygraphviz -E econml --with docs
 
       - name: Run Advanced Tests
         run: poetry run poe test_advanced
