+++ dowhy/utils/regression.py
@@ -1,10 +1,4 @@
 import numpy as np
-from econml.sklearn_extensions.model_selection import GridSearchCVList
-from sklearn.compose import ColumnTransformer
-from sklearn.ensemble import GradientBoostingRegressor, RandomForestRegressor
-from sklearn.linear_model import Lasso
-from sklearn.pipeline import Pipeline
-from sklearn.preprocessing import StandardScaler
 
 
 def get_numeric_features(X):
@@ -21,54 +15,6 @@ def get_numeric_features(X):
         col_index = X.columns.get_loc(col_name)
         numeric_features.append(col_index)
     return numeric_features
-
-
-def get_generic_regressor(
-    cv, X, Y, max_degree=3, estimator_list=None, estimator_param_list=None, numeric_features=None
-):
-    """
-    Finds the best estimator for regression function (g_s)
-
-    :param cv: training and testing data indices obtained afteer Kfolding the dataset
-    :param X: regressors data for training the regression model
-    :param Y: outcome data for training the regression model
-    :param max_degree: degree of the polynomial function used to approximate the regression function
-    :param estimator_list: list of estimator objects for finding the regression function
-    :param estimator_param_list: list of dictionaries with parameters for tuning respective estimators in estimator_list
-    :param numeric_features: list of indices of numeric features in the dataset
-
-    :returns: estimator for Reisz Regression function
-    """
-    if estimator_list is not None:
-        estimator = GridSearchCVList(
-            estimator_list, estimator_param_list, cv=cv, scoring="neg_mean_squared_error", n_jobs=-1
-        ).fit(X, Y)
-        return estimator.best_estimator_
-    else:
-        estimator = GridSearchCVList(
-            [
-                RandomForestRegressor(n_estimators=100, random_state=120),
-                Pipeline(
-                    [
-                        (
-                            "scale",
-                            ColumnTransformer([("num", StandardScaler(), numeric_features)], remainder="passthrough"),
-                        ),
-                        ("lasso_model", Lasso()),
-                    ]
-                ),
-                GradientBoostingRegressor(),
-            ],
-            param_grid_list=[
-                {"n_estimators": [50], "max_depth": [3, 4, 5], "min_samples_leaf": [10, 50]},
-                {"lasso_model__alpha": [0.01, 0.001, 1e-4, 1e-5, 1e-6]},
-                {"learning_rate": [0.01, 0.001], "n_estimators": [50, 200]},
-            ],
-            cv=cv,
-            scoring="neg_mean_squared_error",
-            n_jobs=-1,
-        ).fit(X, Y)
-        return estimator.best_estimator_
 
 
 def generate_moment_function(W, g):
