+++ tests/causal_refuters/test_add_unobserved_common_cause.py
@@ -303,6 +303,7 @@ class TestAddUnobservedCommonCauseRefuter(object):
         assert refute2.stats["robustness_value"] >= rvalue_threshold and refute2.stats["robustness_value"] <= 1
         assert mock_fig.call_count > 0  # we patched figure plotting call to avoid drawing plots during tests
 
+    @pytest.mark.econml
     @pytest.mark.parametrize(
         ["estimator_method", "effect_fraction_on_treatment", "benchmark_common_causes", "simulation_method"],
         [
@@ -391,6 +392,7 @@ class TestAddUnobservedCommonCauseRefuter(object):
         # we patched figure plotting call to avoid drawing plots during tests
         assert mock_fig.call_count > 0
 
+    @pytest.mark.econml
     @pytest.mark.parametrize(
         ["estimator_method", "effect_fraction_on_outcome", "benchmark_common_causes", "simulation_method"],
         [
